pyproject.toml
[tool.poetry]
name = "need_gas"
version = "0.1.0"
description = ""
authors = ["denis <deniseduardoisidrogonzalez@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
requests = "^2.26.0"
celery = "^5.1.2"
redis = "^3.5.3"
django-cors-headers = "^3.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
